Add files via upload
This is a python program using which we can generate 1000s of ID cards within a couple of seconds and hence it saves a lot of time. It is a better way of approach for making id card. Since, traditional method includes a lot of time in designing each and every element of id cards. It is economical too. Fun